Car Lease system/apps/api/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
asyncpg==0.27.0
databases==0.7.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.24.0
pytest==7.4.0
paypalcheckoutsdk==1.0.0
respx==0.20.0
reportlab==4.0.0
pillow==9.5.0
